Database + popup + clearing ability added

+++ main.py
@@ -2,11 +2,12 @@ from PyQt5 import QtCore, QtGui, QtWidgets
 import sys
 import os
 import shutil
+import pickle
 
 
 class Ui_MainWindow(object):
     def __init__(self):
-        self.dirs = []
+        self.popup_status = False
 
     def setupUi(self, MainWindow):
         MainWindow.setObjectName("MainWindow")
@@ -16,7 +17,7 @@ class Ui_MainWindow(object):
         sizePolicy.setVerticalStretch(0)
         sizePolicy.setHeightForWidth(MainWindow.sizePolicy().hasHeightForWidth())
         MainWindow.setSizePolicy(sizePolicy)
-        MainWindow.setMinimumSize(QtCore.QSize(510, 235))
+        MainWindow.setMinimumSize(QtCore.QSize(800, 400))
         MainWindow.setMaximumSize(QtCore.QSize(100000, 50000))
         font = QtGui.QFont()
         font.setFamily("MS Reference Specialty")
@@ -40,7 +41,6 @@ class Ui_MainWindow(object):
                                  ".QPushButton {\n"
                                  "    border-radius: 7px;\n"
                                  "    background-color: rgb(60, 60, 60);\n"
-                                 "    transition-duration: 1s;\n"
                                  "    padding: 0 5;\n"
                                  "    color: rgb(255, 44, 44);\n"
                                  "}\n"
@@ -97,7 +97,7 @@ class Ui_MainWindow(object):
         self.mainLabel.setMinimumSize(QtCore.QSize(0, 30))
         font = QtGui.QFont()
         font.setFamily("8BIT WONDER")
-        font.setPointSize(12)
+        font.setPointSize(18)
         font.setKerning(True)
         font.setStyleStrategy(QtGui.QFont.PreferDefault)
         self.mainLabel.setFont(font)
@@ -201,62 +201,7 @@ class Ui_MainWindow(object):
         self.formLayout.setContentsMargins(5, 5, 5, 5)
         self.formLayout.setObjectName("formLayout")
 
-        # # xButton_1
-        #
-        # self.xButton_1 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        # sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Preferred, QtWidgets.QSizePolicy.Preferred)
-        # sizePolicy.setHorizontalStretch(0)
-        # sizePolicy.setVerticalStretch(0)
-        # sizePolicy.setHeightForWidth(self.xButton_1.sizePolicy().hasHeightForWidth())
-        # self.xButton_1.setSizePolicy(sizePolicy)
-        # self.xButton_1.setMinimumSize(QtCore.QSize(0, 0))
-        # font = QtGui.QFont()
-        # font.setPointSize(8)
-        # font.setBold(True)
-        # font.setWeight(75)
-        # self.xButton_1.setFont(font)
-        # self.xButton_1.setObjectName("xButton_1")
-        # self.formLayout.setWidget(0, QtWidgets.QFormLayout.LabelRole, self.xButton_1)
-        #
-        # # checkBox_1
-        #
-        # self.checkBox_1 = QtWidgets.QCheckBox(self.scrollAreaWidgetContents)
-        # sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Preferred)
-        # sizePolicy.setHorizontalStretch(0)
-        # sizePolicy.setVerticalStretch(0)
-        # sizePolicy.setHeightForWidth(self.checkBox_1.sizePolicy().hasHeightForWidth())
-        # self.checkBox_1.setSizePolicy(sizePolicy)
-        # font = QtGui.QFont()
-        # font.setPointSize(7)
-        # self.checkBox_1.setFont(font)
-        # self.checkBox_1.setObjectName("checkBox_1")
-        # self.formLayout.setWidget(0, QtWidgets.QFormLayout.FieldRole, self.checkBox_1)
-        #
-        # # xButton_2
-        #
-        # self.xButton_2 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        # sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Preferred, QtWidgets.QSizePolicy.Preferred)
-        # sizePolicy.setHorizontalStretch(0)
-        # sizePolicy.setVerticalStretch(0)
-        # sizePolicy.setHeightForWidth(self.xButton_2.sizePolicy().hasHeightForWidth())
-        # self.xButton_2.setSizePolicy(sizePolicy)
-        # font = QtGui.QFont()
-        # font.setBold(True)
-        # font.setWeight(75)
-        # self.xButton_2.setFont(font)
-        # self.xButton_2.setObjectName("xButton_2")
-        # self.formLayout.setWidget(1, QtWidgets.QFormLayout.LabelRole, self.xButton_2)
-        #
-        # # checkBox_2
-        #
-        # self.checkBox_2 = QtWidgets.QCheckBox(self.scrollAreaWidgetContents)
-        # font = QtGui.QFont()
-        # font.setPointSize(7)
-        # self.checkBox_2.setFont(font)
-        # self.checkBox_2.setObjectName("checkBox_2")
-        # self.formLayout.setWidget(1, QtWidgets.QFormLayout.FieldRole, self.checkBox_2)
-
-        # ------------------
+        # scrollArea
 
         self.scrollArea.setWidget(self.scrollAreaWidgetContents)
         self.horizontalLayout.addWidget(self.scrollArea)
@@ -294,49 +239,61 @@ class Ui_MainWindow(object):
 
     def retranslateUi(self, MainWindow):
         _translate = QtCore.QCoreApplication.translate
-        MainWindow.setWindowTitle(_translate("MainWindow", "MainWindow"))
+        MainWindow.setWindowTitle(_translate("MainWindow", "Auto Cleaner"))
         self.mainLabel.setText(_translate("MainWindow", "Enter directory for cleaning"))
         self.settingsButton.setText(_translate("MainWindow", "⚙"))
         self.addButton.setText(_translate("MainWindow", "ADD"))
-        # self.xButton_1.setText(_translate("MainWindow", "X"))
-        # self.checkBox_1.setText(_translate("MainWindow", "dsfsdf"))
-        # self.xButton_2.setText(_translate("MainWindow", "X"))
-        # self.checkBox_2.setText(_translate("MainWindow", "CheckBox"))
         self.clearButton.setText(_translate("MainWindow", "START"))
         self.actionSettings.setText(_translate("MainWindow", "Settings"))
 
     def clear_button_clicked(self):
-        path = self.dirLineEdit.text()
 
-        if os.path.isdir(path):
-            files_deleted = 0
-            files_amount = len(os.listdir(path))
-            print('Очистка ' + str(files_amount) + ' файлов в дериктории:\n' + path)
-
-            files = os.listdir(path)
-            for file in files:
-                try:
-                    file_path = os.path.join(path, file)
-                    if os.path.isfile(file_path):
-                        os.remove(file_path)
-                        files_deleted += 1
-                    else:
-                        shutil.rmtree(file_path)
-                except PermissionError:
-                    pass
-
-            print('\nУдаленно          ' + str(files_deleted) + ' файлов\nПроигнорированно  ' + str(
-                files_amount - files_deleted) + ' файлов')
-            self.mainLabel.setText('Done')
-
+        dirs = self.get_dirs()
+        if not dirs:
+            self.mainLabel.setText('No paths added')
         else:
-            self.mainLabel.setText('Invalid Path')
+            self.show_popup()
+            if self.popup_status:
+                files_deleted = 0
+                files_size = 0
+                for path in dirs:
+                    if os.path.isdir(path):
+                        files = os.listdir(path)
+                        for file in files:
+                            try:
+                                file_path = os.path.join(path, file)
+                                files_size += int(os.path.getsize(file_path))
+                                if os.path.isfile(file_path):
+                                    os.remove(file_path)
+                                    files_deleted += 1
+                                else:
+                                    shutil.rmtree(file_path)
+                            except PermissionError:
+                                pass
+                    else:
+                        self.mainLabel.setText('Invalid Path')
+
+                # Label changing
+
+                if files_deleted == 0:
+                    self.mainLabel.setText('No files deleted')
+                elif files_deleted == 1:
+                    self.mainLabel.setText(str(files_deleted) + ' file deleted (' +
+                                           self.get_file_size(files_size) + ')')
+                else:
+                    self.mainLabel.setText(str(files_deleted) + ' files deleted ( ' +
+                                           self.get_file_size(files_size) + ' )')
+            else:
+                self.mainLabel.setText("Enter directory for cleaning")
 
     def add_button_clicked(self):
         path = self.dirLineEdit.text()
 
         if os.path.isdir(path):
-            self.dirs.append(path)
+            dirs = self.get_dirs()
+            dirs.append(path)
+            self.load_dirs(dirs)
+
             self.mainLabel.setText("Enter directory for cleaning")
             self.update_scroll_area()
         else:
@@ -347,13 +304,10 @@ class Ui_MainWindow(object):
 
         while self.formLayout.count() != 0:
             self.formLayout.removeRow(0)
-            print(self.formLayout.count())
 
-        for n in range(0, len(self.dirs)):
+        for n in range(0, len(self.get_dirs())):
 
             # xButton
-
-            x_button = "xButton_" + str(n)
 
             self.x_button = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
 
@@ -368,7 +322,7 @@ class Ui_MainWindow(object):
             font.setBold(True)
             font.setWeight(75)
             self.x_button.setFont(font)
-            self.x_button.setObjectName(x_button)
+            self.x_button.setObjectName('xButton_' + str(n))
             self.x_button.setText("X")
             self.x_button.clicked.connect(lambda: self.remove_dir(n))
 
@@ -376,29 +330,60 @@ class Ui_MainWindow(object):
 
             # checkBox
 
-            check_box = "checkBox_" + str(n)
-
-            self.check_box = QtWidgets.QCheckBox(self.scrollAreaWidgetContents)
-            sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Preferred)
-            sizePolicy.setHorizontalStretch(0)
-            sizePolicy.setVerticalStretch(0)
-            sizePolicy.setHeightForWidth(self.check_box.sizePolicy().hasHeightForWidth())
-            self.check_box.setSizePolicy(sizePolicy)
-            font = QtGui.QFont()
+            self.check_box = QtWidgets.QLabel(self.scrollAreaWidgetContents)
             font.setPointSize(7)
             self.check_box.setFont(font)
-            self.check_box.setObjectName(check_box)
-            self.check_box.setText(self.dirs[n])
+            self.check_box.setObjectName('checkBox_' + str(n))
+            self.check_box.setText(self.get_dirs()[n])
 
             self.formLayout.setWidget(n, QtWidgets.QFormLayout.FieldRole, self.check_box)
 
-        print(self.formLayout.itemAt(0))
-
     def remove_dir(self, n):
-        self.dirs.pop(n)
+        dirs = self.get_dirs()
+        dirs.pop(n)
+        self.load_dirs(dirs)
         self.update_scroll_area()
 
+    def get_dirs(self):
+        with open('user.data', 'rb') as file:
+            return pickle.load(file)
 
+    def load_dirs(self, dirs):
+        with open('user.data', 'wb') as file:
+            pickle.dump(dirs, file)
+
+    def reset_dirs(self):
+        with open('user.data', 'wb') as file:
+            pickle.dump([], file)
+
+    def get_file_size(self, file_size):
+        file_size = file_size/1024
+        ending = 'KB'
+        if file_size >= 1024:
+            file_size /= 1024
+            ending = 'MB'
+        if file_size >= 1024:
+            file_size /= 1024
+            ending = 'GB'
+        file_size = round(file_size, 2)
+
+        return str(file_size) + ' ' + ending
+
+    def show_popup(self):
+        popup = QtWidgets.QMessageBox()
+        popup.setWindowTitle('Warning!')
+        popup.setText('Do you really want to start?')
+        popup.setIcon(QtWidgets.QMessageBox.Warning)
+        popup.setStandardButtons(QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.No)
+        popup.setDefaultButton(QtWidgets.QMessageBox.No)
+        popup.buttonClicked.connect(self.popup_button_pressed)
+        popup.exec()
+
+    def popup_button_pressed(self, button):
+        if button.text() == '&No':
+            self.popup_status = False
+        else:
+            self.popup_status = True
 
 
 if __name__ == "__main__":
@@ -407,5 +392,6 @@ if __name__ == "__main__":
     MainWindow = QtWidgets.QMainWindow()
     ui = Ui_MainWindow()
     ui.setupUi(MainWindow)
+    ui.update_scroll_area()
     MainWindow.show()
     sys.exit(app.exec_())
